Revert "änderungen an models und Dokumentation hinzugefügt"
This reverts commit 06bda48e82783737c50cdc0acb5c2f57064c0540.

+++ carSharingNow/carSharingNow.zip_expanded/carSharingNow/src/main/java/com/csn/carSharingNow/controller/CarController.java
@@ -27,7 +27,7 @@ public class CarController {
         return carList;
     }
 
-    public Car getCarByID(Long Id) {
+    public Car getCarByID(int Id) {
         Car car = carRepository.findCarById(Id);
         return car;
     }
